fix(#15) 은선님 피드백 반영, 일부 파일의 패키지 이동
