+++ src/main/java/com/limito/payment/presentation/dto/response/FailLogPaymentResponseV1.java
@@ -1,0 +1,26 @@
+package com.limito.payment.presentation.dto.response;
+
+import com.limito.payment.domain.enums.PaymentStatusEnum;
+import com.limito.payment.domain.enums.RefundStatusEnum;
+import lombok.*;
+
+@ToString
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class FailLogPaymentResponseV1 {
+    private String paymentId;
+    private PaymentStatusEnum paymentStatus;
+    private RefundStatusEnum refundStatus;
+    private String pgTransactionId;
+    private String pgProvider;
+    private int retryCount;
+    private String failureReason;
+    private String pgErrorCode;
+    private String pgErrorMessage;
+    private Integer httpStatus;
+    private String apiEndpoint;
+    private String requestPayload;
+    private String responsePayload;
+}
