+++ src/main/resources/application.yaml
@@ -15,6 +15,9 @@ spring:
         dialect: org.hibernate.dialect.PostgreSQLDialect
         format_sql: true
         show_sql: false
+        type:
+          descriptor:
+            sql: trace
     open-in-view: false
   thymeleaf:
     prefix: classpath:/templates/
